Data for PARAMETER_NAME, PARAMETER_TYPE, etc. in the Parameters table

Returns comprehensive parameter data combining quality and chemical parameters.

Usage

parameters_vocabulary()

Value

A data frame combining quality and chemical parameter options

Details

Combines quality parameters with chemical parameters from the ChemOnt chemical taxonomy (Djoumbou Feunang et al. 2016, doi:10.1186/s13321-016-0174-y ). Quality parameters are read from quality_parameters.parquet and chemical parameters from ClassyFire_Taxonomy_2025_02.parquet. The resulting dataset includes columns for parameter classification, chemical identifiers (InChIKey, PubChem CID, CAS RN), and measurement types.

Note: Quality parameters have not yet been comprehensively validated as we are not aware of any taxonomy to validate them against.

Data sources:

extdata/quality_parameters.parquet

extdata/ClassyFire_Taxonomy_2025_02.parquet

See also

Examples

head(parameters_vocabulary(), 10)
#> # A tibble: 10 × 9
#>    PARAMETER_TYPE    PARAMETER_NAME      PARAMETER_TYPE_SUB MEASURED_TYPE CAS_RN
#>    <chr>             <chr>               <chr>              <chr>         <chr> 
#>  1 Quality parameter pH                  pH                 Physical par… NA    
#>  2 Quality parameter Dissolved oxygen    Dissolved oxygen   Concentration 7782-…
#>  3 Quality parameter Temperature         Temperature        Physical par… NA    
#>  4 Quality parameter Conductivity        Conductivity       Physical par… NA    
#>  5 Quality parameter Salinity            Salinity           Physical par… NA    
#>  6 Quality parameter Turbidity           Turbidity          Physical par… NA    
#>  7 Quality parameter Total organic carb… Total organic car… Concentration NA    
#>  8 Quality parameter Alkalinity          Other              Concentration NA    
#>  9 Quality parameter Hardness            Other              Concentration NA    
#> 10 Quality parameter Chlorophyll a       Other              Concentration 479-6…
#> # ℹ 4 more variables: PUBCHEM_CID <int>, INCHIKEY_SD <chr>, ENTERED_BY <chr>,
#> #   PARAMETER_NAME_SUB <chr>