Analytical protocols controlled vocabulary

Returns analytical protocol options with short and long names.

Usage

analytical_protocols_vocabulary()

Value

A tibble with Protocol_Type, Short_Name, and Long_Name columns

Details

Provides standardised analytical method classifications as a tibble with three columns:

Protocol_Type: Always "Analytical Protocol"

Short_Name: Abbreviated protocol name (e.g., "GC-MS", "LC-MS/MS", "ICP-MS")

Long_Name: Full descriptive name (e.g., "Gas chromatography mass spectrometry", "Liquid chromatography tandem mass spectrometry", "Inductively coupled plasma mass spectrometry")

Available protocols include: Not relevant, Not reported, GC-MS, LC-MS, LC-MS/MS, GC-MS/MS, UPLC, ICP-MS, ICP-OES, AAS, XRF, Ion chromatography, Spectrophotometry, Fluorescence, Other

See also